NF: extract DirectoryValidator in its own class
